backend/requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
flask-socketio==5.3.6
flask-limiter==3.5.0

# Database
psycopg2-binary==2.9.7
redis==5.0.1
sqlalchemy==2.0.23

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
diffusers==0.24.0
transformers==4.35.2
accelerate==0.24.1
safetensors==0.4.1
xformers==0.0.22.post7
controlnet-aux==0.0.7

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4

# HTTP and Async
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Background Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# File Storage
boto3==1.34.0
botocore==1.34.0

# Validation and Serialization
marshmallow==3.20.1
pydantic==2.5.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
python-socketio==5.10.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
bcrypt==4.1.2
cryptography==41.0.8
pyjwt==2.8.0

# Image Enhancement
rembg==2.0.50
gfpgan==1.3.8
